Moved test up a level for fixed imports
