+++ ee/codegen/src/__test__/workflow.test.ts
@@ -1,12 +1,16 @@
 import { Writer } from "@fern-api/python-ast/core/Writer";
+import { v4 as uuidv4 } from "uuid";
 import { DocumentIndexRead } from "vellum-ai/api";
 import { DocumentIndexes as DocumentIndexesClient } from "vellum-ai/api/resources/documentIndexes/client/Client";
 import { afterEach, vi } from "vitest";
 
 import { workflowContextFactory } from "./helpers";
+import { edgesFactory } from "./helpers/edge-data-factories";
 import { inputVariableContextFactory } from "./helpers/input-variable-context-factory";
 import {
+  conditionalNodeFactory,
   entrypointNodeDataFactory,
+  finalOutputNodeFactory,
   searchNodeDataFactory,
   templatingNodeFactory,
   terminalNodeDataFactory,
@@ -238,6 +242,63 @@ describe("Workflow", () => {
         },
       ];
       workflowContext.addWorkflowEdges(edges);
+
+      const inputs = codegen.inputs({ workflowContext });
+      const workflow = codegen.workflow({
+        moduleName,
+        workflowContext,
+        inputs,
+      });
+
+      workflow.getWorkflowFile().write(writer);
+      expect(await writer.toStringFormatted()).toMatchSnapshot();
+    });
+
+    it("should handle loops of conditionals and import Graph.from_set properly", async () => {
+      const startNode = conditionalNodeFactory({
+        label: "Start Node",
+      });
+      await createNodeContext({
+        workflowContext,
+        nodeData: startNode,
+      });
+
+      const loopCheckNode = conditionalNodeFactory({
+        id: uuidv4(),
+        label: "Loop Check Node",
+        ifSourceHandleId: uuidv4(),
+        elseSourceHandleId: uuidv4(),
+        targetHandleId: uuidv4(),
+      });
+      await createNodeContext({
+        workflowContext,
+        nodeData: loopCheckNode,
+      });
+
+      const finalOutput = finalOutputNodeFactory();
+      await createNodeContext({
+        workflowContext,
+        nodeData: finalOutput,
+      });
+
+      const topNode = templatingNodeFactory({
+        label: "Top Node",
+      });
+      await createNodeContext({
+        workflowContext,
+        nodeData: topNode,
+      });
+
+      workflowContext.addWorkflowEdges(
+        edgesFactory([
+          [entrypointNode, startNode],
+          [[startNode, "0"], topNode],
+          [[startNode, "1"], loopCheckNode],
+          [[loopCheckNode, "0"], startNode],
+          [[loopCheckNode, "1"], finalOutput],
+          [[topNode, "0"], loopCheckNode],
+        ])
+      );
 
       const inputs = codegen.inputs({ workflowContext });
       const workflow = codegen.workflow({
